Add unit test

+++ src/main/kotlin/com/hadi/kotlin/domain/Dtos.kt
@@ -1,10 +1,11 @@
 package com.hadi.kotlin.domain
 
 import com.fasterxml.jackson.annotation.JsonFormat
-import java.math.BigDecimal
 import java.time.LocalDate
 import java.util.*
-import javax.validation.constraints.*
+import javax.validation.constraints.Future
+import javax.validation.constraints.FutureOrPresent
+import javax.validation.constraints.NotNull
 
 data class FindPolicyDto(
   @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
@@ -23,7 +24,7 @@ data class InsuredPersonDto(
   @NotNull val uuid: UUID?,
   @NotNull val firstName: String,
   @NotNull val secondName: String,
-  @NotNull val premium: BigDecimal
+  @NotNull val premium: Double
 )
 
 data class IntegratedPolicyDto(
@@ -31,5 +32,5 @@ data class IntegratedPolicyDto(
   @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
   @NotNull @field:FutureOrPresent val startDate: LocalDate,
   @NotNull val insuredPersons: List<InsuredPersonDto>,
-  @NotNull val totalPremium: BigDecimal?
+  @NotNull val totalPremium: Double?
 )
